+++ src/Veneer/Plugin/Wrapper.php
@@ -14,6 +14,7 @@ use Closure;
 use DecodeLabs\Exceptional;
 use DecodeLabs\Glitch\Dumpable;
 use IteratorAggregate;
+use Stringable;
 use Traversable;
 
 /**
@@ -25,6 +26,7 @@ use Traversable;
 class Wrapper implements
     ArrayAccess,
     IteratorAggregate,
+    Stringable,
     Dumpable
 {
     /**
@@ -100,7 +102,7 @@ class Wrapper implements
 
         if (!$this->plugin instanceof ArrayAccess) {
             throw Exceptional::Runtime(
-                'Plugin does not implement ArrayAccess'
+                message: 'Plugin does not implement ArrayAccess'
             );
         }
 
@@ -119,7 +121,7 @@ class Wrapper implements
 
         if (!$this->plugin instanceof ArrayAccess) {
             throw Exceptional::Runtime(
-                'Plugin does not implement ArrayAccess'
+                message: 'Plugin does not implement ArrayAccess'
             );
         }
 
@@ -138,7 +140,7 @@ class Wrapper implements
 
         if (!$this->plugin instanceof ArrayAccess) {
             throw Exceptional::Runtime(
-                'Plugin does not implement ArrayAccess'
+                message: 'Plugin does not implement ArrayAccess'
             );
         }
 
@@ -157,7 +159,7 @@ class Wrapper implements
 
         if (!$this->plugin instanceof ArrayAccess) {
             throw Exceptional::Runtime(
-                'Plugin does not implement ArrayAccess'
+                message: 'Plugin does not implement ArrayAccess'
             );
         }
 
@@ -180,13 +182,30 @@ class Wrapper implements
 
         if (!$this->plugin instanceof Traversable) {
             throw Exceptional::Runtime(
-                'Plugin does not implement Traversable'
+                message: 'Plugin does not implement Traversable'
             );
         }
 
         return $this->plugin;
     }
 
+    /**
+     * Convert to string
+     */
+    public function __toString(): string
+    {
+        if($this->plugin === null) {
+            $this->getVeneerPlugin();
+        }
+
+        if(!$this->plugin instanceof Stringable) {
+            throw Exceptional::Runtime(
+                message: 'Plugin does not implement Stringable'
+            );
+        }
+
+        return $this->plugin->__toString();
+    }
 
     /**
      * Export for dump inspection