+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.12.3 (2025-02-14)
+* Added Stringable to Wrapper
+* Removed Pandora dev dependency
+* Improved Exception syntax
+* Updated dependencies
+
 ## v0.12.2 (2025-02-12)
 * Removed Coercion dependency